requirements.txt
Django==4.2.5
djangorestframework==3.14.0
mysql-connector-python==8.0.33